rename mill.worker to compilerworker
