+++ runner/meta/package.mill
@@ -6,7 +6,7 @@ import mill.contrib.buildinfo.BuildInfo
 object `package` extends build.MillPublishScalaModule with BuildInfo {
   def moduleDeps = Seq(
     build.runner.codesig,
-    build.runner.worker.api,
+    build.runner.compilerworker.api,
     build.libs.main
   )
 
@@ -15,7 +15,7 @@ object `package` extends build.MillPublishScalaModule with BuildInfo {
   def buildInfoMembers = Seq(
     BuildInfo.Value(
       "bootstrapDeps",
-      build.runner.worker.bootstrapDeps().mkString(";"),
+      build.runner.compilerworker.bootstrapDeps().mkString(";"),
       "Depedendencies used to bootstrap the scala compiler worker."
     )
   )
