+++ runner/meta/src/mill/meta/FileImportGraph.scala
@@ -3,7 +3,7 @@ package mill.meta
 import mill.api.internal.internal
 import mill.constants.CodeGenConstants.*
 import mill.constants.OutFiles.*
-import mill.worker.api.MillScalaParser
+import mill.compilerworker.api.MillScalaParser
 import scala.collection.mutable
 import scala.jdk.CollectionConverters.CollectionHasAsScala
 import mill.internal.Util.backtickWrap
